+++ src/app/dashboard/page.tsx
@@ -15,7 +15,7 @@ import {
   Square,
   HStack,
 } from '@chakra-ui/react'
-import { BraindaoLogo3 } from '@/components/braindao-logo-3'
+import { BraindaoLogo3 } from '@/components/icons/braindao-logo-3'
 import { Dict } from '@chakra-ui/utils'
 import {
   CUSTOM_GRAPH_PERIODS,
@@ -68,7 +68,7 @@ const Home: NextPage = () => {
   })
   const { startDate, endDate } = getDateRange(stakeValue as string)
   const { data } = useGetStakeValueQuery({ startDate, endDate })
-  const stakeGraphData = data?.map((dt) => ({
+  const stakeGraphData = data?.map(dt => ({
     amt: parseFloat(dt.amount),
     name: new Date(dt.created).toISOString().slice(0, 10),
   }))
@@ -136,7 +136,7 @@ const Home: NextPage = () => {
         })
       })
 
-      Promise.resolve(res2).then((data) => {
+      Promise.resolve(res2).then(data => {
         setMarketData(data)
       })
     }
@@ -218,7 +218,7 @@ const Home: NextPage = () => {
               graphTitle="IQ price"
               height={120}
             >
-              {GRAPH_PERIODS.map((btn) => {
+              {GRAPH_PERIODS.map(btn => {
                 return (
                   <GraphPeriodButton
                     key={btn.period}
@@ -238,7 +238,7 @@ const Home: NextPage = () => {
               graphTitle="IQ Staked Over time"
               height={200}
             >
-              {CUSTOM_GRAPH_PERIODS.map((btn) => {
+              {CUSTOM_GRAPH_PERIODS.map(btn => {
                 return (
                   <GraphPeriodButton
                     key={btn.period}
@@ -295,7 +295,7 @@ const Home: NextPage = () => {
               />
               <Box mt={{ lg: '2', '2xl': '-11' }}>
                 <Flex w="full" direction="column" gap={{ base: 2, md: 4 }}>
-                  {holders.map((item) => (
+                  {holders.map(item => (
                     <HStack w="full">
                       <Square
                         bg={
