+++ src/components/lock/StakeIQ.tsx
@@ -1,4 +1,4 @@
-import { BraindaoLogo3 } from '@/components/braindao-logo-3'
+import { BraindaoLogo3 } from '@/components/icons/braindao-logo-3'
 import {
   calculateReturn,
   getValueFromBigInt,
@@ -251,7 +251,7 @@ const StakeIQ = ({ exchangeRate }: { exchangeRate: number }) => {
         <Flex align="center" gap="2.5" w="full">
           <Input
             variant="unstyled"
-            onChange={(e) => updateIqToBeLocked(e.target.value)}
+            onChange={e => updateIqToBeLocked(e.target.value)}
             placeholder="23.00"
             value={userInput}
             color="fadedText4"
@@ -277,7 +277,7 @@ const StakeIQ = ({ exchangeRate }: { exchangeRate: number }) => {
         </Flex>
       </VStack>
       {userTotalIQLocked < 1 && (
-        <LockSlider updateLockend={(newDate) => updateLockend(newDate)} />
+        <LockSlider updateLockend={newDate => updateLockend(newDate)} />
       )}
       <IconButton
         icon={<RiArrowDownLine />}
