+++ src/components/vote/empty-state.tsx
