--- src/data/LockData.ts
@@ -1,18 +1,0 @@
-export const LOCK_DATA = [
-  {
-    title: 'Total supply of HiIQ',
-    value: '1,443,751,656 HIIQ',
-  },
-  {
-    title: 'Total volume locked',
-    value: '409,581,181 IQ',
-  },
-  {
-    title: 'Yield earned over 4years',
-    value: '404.45%',
-  },
-  {
-    title: 'Annual percentage rate',
-    value: '44.18%',
-  },
-]
