+++ src/data/SidebarData.ts
@@ -1,4 +1,5 @@
-import { BraindaoLogo } from '@/components/braindao-logo'
+
+import { BraindaoLogo } from '@/components/icons/braindao-logo'
 import { IconProps } from '@chakra-ui/react'
 import { IconType } from 'react-icons/lib'
 import {
